ai-service/requirements.txt
annotated-types==0.5.0
audeer==1.20.1
audioread==3.0.0
audobject==0.7.9
audonnx==0.6.3
av==10.0.0
backports.functools-lru-cache==1.6.6
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.4.1
cn2an==0.5.22
coloredlogs==15.0.1
ctranslate2==3.19.0
Cython==3.0.2
decorator==5.1.1
eng-to-ipa==0.0.2
faster-whisper==0.8.0
filelock==3.12.3
flatbuffers==23.5.26
fsspec==2023.9.0
grpcio==1.58.0
grpcio-reflection==1.58.0
grpcio-tools==1.58.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
indic-transliteration==2.3.49
inflect==7.0.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
ko-pron==1.3
lazy_loader==0.3
librosa==0.10.1
lit==16.0.6
llvmlite==0.40.1
MarkupSafe==2.1.3
mecab-python3==1.0.6
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
num-thai==0.0.5
numba==0.57.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnx==1.14.1
onnxruntime==1.15.1
OpenCC==1.1.6
openjtalk==0.3.0.dev2
opus-fast-mosestokenizer==0.0.8.3
oyaml==1.0
packaging==23.1
plac==1.3.5
platformdirs==3.10.0
pooch==1.7.0
proces==0.1.7
protobuf==4.24.2
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
PyOgg==0.6.14a1
pypinyin==0.49.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
roman==4.1
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
soundfile==0.12.1
soxr==0.3.6
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
torch==2.0.1
tqdm==4.66.1
triton==2.0.0
typer==0.9.0
typing_extensions==4.7.1
Unidecode==1.3.6
unidic==1.1.0
urllib3==2.0.4
wasabi==0.10.1
zipp==3.16.2
